Move signOut dialog to a separate file
